prevent overlapping writes
If it was ever possible to run the shadow algorithm on multiple threads,
it would be very nice if each octal block mutated a unique collection of
mask cells; this would allow for simpler (and faster!) parallel-ization.

Previously, there was some overlap due to the algorithm used; each octal needs
to observe its edges in order to iterate correctly. Technically, this overlap
would *not* result in a race condition; the final image will be the same
irrespective of the order of the octals rendered ... but theoretically it is
nicer this way.